complete smart search

+++ src/main/java/high/skill/girl/project/calories_counter/dto/ProductDto.java
@@ -2,9 +2,11 @@ package high.skill.girl.project.calories_counter.dto;
 
 import io.micronaut.serde.annotation.Serdeable;
 
+import java.math.BigDecimal;
+
 @Serdeable
 public record ProductDto(String name,
                          int calories,
-                         float proteins,
-                         float fats,
-                         float carbohydrates) { }
+                         BigDecimal proteins,
+                         BigDecimal fats,
+                         BigDecimal carbohydrates) { }
